Cultureinfo working. Average distance/duration working.

+++ MyBikeApp/Controllers/StationsController.cs
@@ -1,20 +1,27 @@
 ﻿using System;
 using System.Collections.Generic;
+
+using System.Globalization;
 using System.Linq;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using MyBikeApp.Data;
 using MyBikeApp.Models;
+using Newtonsoft.Json.Linq;
 using X.PagedList;
+using static System.Collections.Specialized.BitVector32;
+using static System.Net.Mime.MediaTypeNames;
 
 namespace MyBikeApp.Controllers
 {
     public class StationsController : Controller
     {
 		private StationsController _Instance = null; public StationsController Instance { get { return _Instance; } set { _Instance = value; } }
-
+        
 
 		private readonly ApplicationDbContext _context;
 		String[] csvLines = System.IO.File.ReadAllLines(@"C:\Users\Omistaja\source\repos\MyBikeApp\MyBikeApp\csv\Helsingin_ja_Espoon_kaupunkipyöräasemat_avoin.csv");
@@ -55,8 +62,8 @@ namespace MyBikeApp.Controllers
 		{
 	//		int beforeCount = _context.Station.Count();
 	//		Console.WriteLine("before clear:" + beforeCount);
-
-			_context.Station.RemoveRange(_context.Station.Where(x => x.Name != ""));
+            if(_context.Station != null)
+			   _context.Station.RemoveRange(_context.Station.Where(x => x.Name != ""));
 			_context.SaveChanges();
 
 		//	int afterCount = _context.Station.Count();
@@ -66,7 +73,6 @@ namespace MyBikeApp.Controllers
 
 		public ViewResult Index(string sortOrder, string currentFilter, string searchString, int? page)
 		{
-
 			ViewBag.CurrentSort = sortOrder;
 			ViewBag.NameSortParm = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
 			ViewBag.AddressSortParm = String.IsNullOrEmpty(sortOrder) ? "return_name_desc" : "";
@@ -125,8 +131,6 @@ namespace MyBikeApp.Controllers
                 return NotFound();
             }
 
-
-
             var station = await _context.Station
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (station == null)
@@ -134,29 +138,119 @@ namespace MyBikeApp.Controllers
                 return NotFound();
             }
 
-            var result = _context.Journey
-              .Where(d => d.DepartureStationName.Contains(station.Name))
-              .GroupBy(q => q.ReturnStationName)
-              .OrderByDescending(g => g.Count())
-              .Take(5)
-              .Select(g => g.Key)
-              .ToList();
+            SharedViewModel StationInfo = new SharedViewModel();
 
-            List<MyBikeApp.Models.Journey> newList = new List<MyBikeApp.Models.Journey>();
-            foreach (var item in result)
-            {
-                MyBikeApp.Models.Journey listItem = new MyBikeApp.Models.Journey();
-                listItem.ReturnStationName = item;
-                newList.Add(listItem);
-            }
-            return View("Details", newList);
+            GetTopFiveDestinations(StationInfo, station);
+			GetTopFiveDepartureStations(StationInfo, station);
+			GetAverageDistances(StationInfo, station);
+			GetGoogleMapsCoordinates(StationInfo, station);
+			
+            return View("Details", StationInfo);
+		}
+		public void GetGoogleMapsCoordinates(SharedViewModel StationInfo, Station station)
+		{
+			var _lat = _context.Station
+				 .Where(d => d.Name.Contains(station.Name)).Select(g => g.Lat).FirstOrDefault();
+			var _lng = _context.Station
+				 .Where(d => d.Name.Contains(station.Name)).Select(g => g.Lon).FirstOrDefault();
+			var culture = CorrectNumberFormat();
 
+			string temp_lat = _lat.ToString();
+			string temp_lng = _lng.ToString();
+			// Console.WriteLine(temp_lat);
+			// Console.WriteLine(temp_lng);
 
-         //   return View(station);
-        }
+			if (!double.TryParse(temp_lat, System.Globalization.NumberStyles.Any, culture, out StationInfo.lat))
+				Console.WriteLine("Error parsin lat");
+			if (!double.TryParse(temp_lng, System.Globalization.NumberStyles.Any, culture, out StationInfo.lng))
+				Console.WriteLine("Error parsin long");
+			//Console.Write("lat: " + StationInfo.lat + " lng: " + StationInfo.lng);
+		}
+		public void GetAverageDistances(SharedViewModel StationInfo, Station station)
+        {
+			StationInfo.TripsFromThisStation = _context.Journey
+			  .Where(d => d.DepartureStationName.Contains(station.Name))
+			  .ToList()
+			  .Count;
 
-        // GET: Stations/Create
-        public IActionResult Create()
+			StationInfo.TripsToThisStation = _context.Journey
+		   .Where(d => d.ReturnStationName.Contains(station.Name))
+		   .ToList()
+		   .Count;
+
+			StationInfo.AverageDistanceFromStation = ((int)_context.Journey
+			  .Where(d => d.DepartureStationName.Contains(station.Name))
+			  .Select(d => d.CoveredDistanceM).Sum());
+			if(StationInfo.TripsFromThisStation > 0)
+				StationInfo.AverageDistanceFromStation = StationInfo.AverageDistanceFromStation / StationInfo.TripsFromThisStation;
+			//Console.Write("AverageDistanceFromStation: " + StationInfo.AverageDistanceFromStation);
+
+			StationInfo.AverageDistanceToStation = ((int)_context.Journey
+			  .Where(d => d.ReturnStationName.Contains(station.Name))
+			  .Select(d => d.CoveredDistanceM).Sum());
+			if(StationInfo.TripsToThisStation > 0)
+				StationInfo.AverageDistanceToStation = StationInfo.AverageDistanceToStation / StationInfo.TripsToThisStation;
+			//Console.Write("AverageDistanceToStation: " + StationInfo.AverageDistanceToStation);
+		}
+
+		public void GetTopFiveDestinations(SharedViewModel StationInfo, Station station)
+		{
+			List<Journey> newList = new List<Journey>();
+			var result = _context.Journey
+			  .Where(d => d.DepartureStationName.Contains(station.Name))
+			  .GroupBy(q => q.ReturnStationName)
+			  .OrderByDescending(g => g.Count())
+			  .Take(5)
+			  .Select(g => g.Key)
+			  .ToList();
+
+			foreach (var item in result)
+			{
+				MyBikeApp.Models.Journey listItem = new MyBikeApp.Models.Journey();
+				listItem.ReturnStationName = item;
+				StationInfo.JourneysFromStation.Add(listItem);
+			}
+		}
+        public void GetTopFiveDepartureStations(SharedViewModel StationInfo, Station station)
+        {
+			var result = _context.Journey
+			  .Where(d => d.ReturnStationName.Contains(station.Name))
+			  .GroupBy(q => q.DepartureStationName)
+			  .OrderByDescending(g => g.Count())
+			  .Take(5)
+			  .Select(g => g.Key)
+			  .ToList();
+
+			foreach (var item in result)
+			{
+				MyBikeApp.Models.Journey listItem = new MyBikeApp.Models.Journey();
+				listItem.DepartureStationName = item;
+				StationInfo.JourneysToStation.Add(listItem);
+			}
+		}
+
+		
+       
+
+		public static CultureInfo CorrectNumberFormat()
+		{
+			var culture = CultureInfo.CreateSpecificCulture(CultureInfo.CurrentCulture.Name);
+			if (culture.NumberFormat.NumberDecimalSeparator != ".")
+			{
+				culture.NumberFormat.NumberDecimalSeparator = ".";
+				culture.NumberFormat.CurrencyDecimalSeparator = ".";
+				culture.NumberFormat.PercentDecimalSeparator = ".";
+				CultureInfo.DefaultThreadCurrentCulture = culture;
+				CultureInfo.DefaultThreadCurrentUICulture = culture;
+
+				Thread.CurrentThread.CurrentCulture = culture;
+				Thread.CurrentThread.CurrentUICulture = culture;
+			}
+            return culture;
+		}
+
+		// GET: Stations/Create
+		public IActionResult Create()
         {
             return View();
         }
